boe_chain/requirements.txt
langchain
chromadb
langchain-community
langchain-chroma
openai
llama-cpp-python
langchain-openai